+++ testdata/full_compute_instance.json
