+++ testdata/full_compute_instance.tf
