+++ tfplan2cai/tfplan_to_cai_test.go
@@ -1,11 +1,10 @@
-package caitest
+// Package tfplan2cai converts tfplan to CAI assets.
+package tfplan2cai
 
 import (
 	"context"
 	"io/ioutil"
 	"testing"
-
-	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/conversion"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -13,8 +12,8 @@ import (
 func TestTFPlanToCAI_noPlanJSON(t *testing.T) {
 	ctx := context.Background()
 	jsonPlan := []byte{}
-	options := &conversion.TFPlanToCAIOptions{}
-	assets, err := conversion.TFPlanToCAI(ctx, jsonPlan, options)
+	options := &TFPlanToCAIOptions{}
+	assets, err := TFPlanToCAI(ctx, jsonPlan, options)
 	assert.Empty(t, assets)
 	assert.Error(t, err)
 }
@@ -26,8 +25,8 @@ func TestTFPlanToCAI_noResourceChanges(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Error parsing %s: %s", f, err)
 	}
-	options := &conversion.TFPlanToCAIOptions{}
-	assets, err := conversion.TFPlanToCAI(ctx, jsonPlan, options)
+	options := &TFPlanToCAIOptions{}
+	assets, err := TFPlanToCAI(ctx, jsonPlan, options)
 	assert.Empty(t, assets)
 	assert.Empty(t, err)
 }
